+++ src/extension.ts
@@ -1,9 +1,10 @@
 'use strict';
 import { ExtensionContext } from 'vscode';
+import { Keyboard } from './commands';
+import { ShowRelatedCommand } from './commands';
 import { FindRelatedApi } from './extensionApi';
 import { Logger } from './logger';
 import { RulesProvider } from './rulesProvider';
-import { ShowRelatedCommand } from './showRelatedCommand';
 
 // this method is called when your extension is activated
 export async function activate(context: ExtensionContext) {
@@ -12,6 +13,7 @@ export async function activate(context: ExtensionContext) {
     const rulesProvider = new RulesProvider(context);
     context.subscriptions.push(rulesProvider);
 
+    context.subscriptions.push(new Keyboard());
     context.subscriptions.push(new ShowRelatedCommand(context, rulesProvider));
 
     const api = new FindRelatedApi(context, rulesProvider);